+++ src/app.rs
@@ -1,6 +1,10 @@
 use std::{cell::RefCell, rc::Rc};
 
-use crossbeam_channel::{Receiver, Sender};
+#[cfg(feature = "crossbeam")]
+use crossbeam::channel::{unbounded as channel, Receiver, Sender};
+#[cfg(not(feature = "crossbeam"))]
+use std::sync::mpsc::{channel, Receiver, Sender};
+
 use floem_reactive::WriteSignal;
 use parking_lot::Mutex;
 use raw_window_handle::HasDisplayHandle;
@@ -160,7 +164,8 @@ impl Application {
         crate::app_delegate::set_app_delegate();
 
         let event_loop_proxy = event_loop.create_proxy();
-        let (sender, receiver) = crossbeam_channel::unbounded();
+        let (sender, receiver) = channel();
+
         *EVENT_LOOP_PROXY.lock() = Some((event_loop_proxy.clone(), sender));
         unsafe {
             Clipboard::init(event_loop.display_handle().unwrap().as_raw());